[aot eager] Support input graphs with list arguments
[ghstack-poisoned]
